split Constants so that env-sourced Dataland constants are seperate from hardcoded constants
