+++ dataland-frontend/src/components/general/apiKey/CreateApiKeyCard.vue
@@ -83,7 +83,8 @@ import Calendar from "primevue/calendar";
 import { calculateExpiryDateAsDateString, calculateDaysFromNow } from "@/utils/DataFormatUtils";
 import UserRolesBadges from "@/components/general/apiKey/UserRolesBadges.vue";
 import { assertDefined } from "@/utils/TypeScriptUtils";
-import { MAX_NUMBER_OF_DAYS_SELECTABLE_FOR_API_KEY_VALIDITY, MS_PER_DAY } from "@/utils/Constants";
+import { MS_PER_DAY } from "@/utils/Constants";
+import { MAX_NUMBER_OF_DAYS_SELECTABLE_FOR_API_KEY_VALIDITY } from "@/DatalandSettings";
 
 export default defineComponent({
   name: "CreateApiKeyCard",
