+++ dataland-frontend/src/components/pages/RequestData.vue
@@ -135,8 +135,9 @@ import InfoCard from "@/components/general/InfoCard.vue";
 import ProgressBar from "@/components/general/ProgressBar.vue";
 import { assertDefined } from "@/utils/TypeScriptUtils";
 import { formatBytesUserFriendly, roundNumber } from "@/utils/NumberConversionUtils";
-import { UPLOAD_FILE_SIZE_DISPLAY_DECIMALS, DATA_REQUEST_UPLOAD_MAX_FILE_SIZE_IN_BYTES } from "@/utils/Constants";
+import { UPLOAD_FILE_SIZE_DISPLAY_DECIMALS } from "@/utils/Constants";
 import TheFooter from "@/components/general/TheFooter.vue";
+import { DATA_REQUEST_UPLOAD_MAX_FILE_SIZE_IN_BYTES } from "@/DatalandSettings";
 
 export default defineComponent({
   name: "RequestData",
