+++ dataland-frontend/tests/e2e/specs/landing-page/FooterSection.ts
@@ -3,6 +3,7 @@ import { DataTypeEnum } from "@clients/backend";
 import { getKeycloakToken } from "@e2e/utils/Auth";
 import { reader_name, reader_pw } from "@e2e/utils/Cypress";
 import { checkFooter } from "@sharedUtils/ElementChecks";
+import { ARRAY_OF_FRAMEWORKS_WITH_VIEW_PAGE } from "@/utils/Constants";
 
 describe("As a user, I expect the footer section to be present and contain relevant legal links", () => {
   describe("Checks that the footer section is present on many pages", () => {
@@ -19,8 +20,7 @@ describe("As a user, I expect the footer section to be present and contain relev
       });
     });
 
-    const frameworksToCheck = Object.values(DataTypeEnum);
-    frameworksToCheck.forEach((framework) => {
+    ARRAY_OF_FRAMEWORKS_WITH_VIEW_PAGE.forEach((framework) => {
       it(`Checks that the footer is present on ${framework}`, () => {
         getKeycloakToken(reader_name, reader_pw).then((token) => {
           cy.browserThen(getStoredCompaniesForDataType(token, framework)).then((storedCompanies) => {
